moved admin into HURUmap
