+++ hurumap/admin/__init__.py
