+++ hurumap/admin/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AdminConfig(AppConfig):
-    name = 'admin'
+    name = 'hurumap.admin'
