+++ hurumap/admin/models.py
@@ -4,6 +4,7 @@ from wagtail.wagtailcore import blocks
 from wagtail.wagtailcore.fields import StreamField
 from wagtail.wagtailcore.models import Page
 from wagtail.wagtailimages.blocks import ImageChooserBlock
+from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 from wagtail.wagtailsearch import index
 
 
@@ -17,6 +18,13 @@ class BlogIndexPage(Page):
 class BlogPage(Page):
     date = models.DateField("Post date")
     intro = models.CharField(max_length=250)
+    featured_image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
     body = StreamField([
         ('heading', blocks.CharBlock(classname="full title")),
         ('paragraph', blocks.RichTextBlock()),
@@ -32,4 +40,8 @@ class BlogPage(Page):
         FieldPanel('date'),
         FieldPanel('intro'),
         StreamFieldPanel('body', classname="full"),
+    ]
+
+    promote_panels = [
+        ImageChooserPanel('featured_image'),
     ]
